+++ projects/orbit/src/public-api.ts
@@ -1,1 +1,2 @@
 export * from '@rhaegal222/orbit';
+export { OrbitTopbarComponent } from './components/topbar/topbar.component';